requirements/proceso.txt
gi
1)creacion del entorno

python -m venv 'NOMBRE DEL ENTORNO'

2) pip install django

3) creo el proyecto -   django-admin startproject 'nombre del proyecto'

4)Instalacion de paquetes
	
	pip install unipath
	pip install pillow

5) pip freeze

PARA ISNTALAR TODO DE UNA; CON EL ARCHIVO LOCAL desde la TERMINAL, COLOCO

pip install -r RUTA

Ejempl

pip install -r requirements/local.txt


PARTE 2

creacion de la base de datos desde postgres sql

postgres-# CREATE DATABASE nombre de la base;

\c dbuser    //para ingresar a la base de datos

CREACION DEL PASSWORD DE USUARIO DE LA BASE DE DATOS

dbuser=# ALTER ROLE grillo WITH PASSWORD 'grillo1974';

una vez comfigurado todo tmb hay que agregar el .local en manage.py // wsgi.py // asgi.py


//// una vez que genere el archivo .gitignore puedo subir todo github

git add .

git commit -m "configuracion del proyecto"

git push


////////////// CREACION DE APPLICATIONS EN DJANGO
django-admin startapp nombre de la aplicaion

/////////////Generacion del super usuario
python manage.py createsuperuser







